Fix integer truncation bug in vint_to_u64()
Commit UUID: a8773110-2d4a-4035-8372-91f9d027199b

+++ trunk/vint.c
@@ -108,7 +108,7 @@ vint_to_u64(const char *x, uint64_t *y, size_t *sz)
     i = len;
     while (i > 0) {
         ++x;
-        ret |= (unsigned char)*x << --i * CHAR_BIT;
+        ret |= (uint64_t)(unsigned char)*x << --i * CHAR_BIT;
     }
 
     *y = ret;
